Frontend: Adicionada a funcao de cadastrar, excluir e editar veiculos.

+++ Sistema-Receptivo/src/main/java/br/com/sosviale/repository/VeiculoRepository.java
@@ -36,7 +36,7 @@ public class VeiculoRepository {
     }
 
     // busca um veículo pelo ID; retorna null se não encontrado
-    public Veiculo buscarPorId(Long id) {
+    public Veiculo buscarPorId(Integer id) {
         if (id == null || id <= 0) throw new IllegalArgumentException("ID inválido.");
         EntityManager em = JPAUtil.getEntityManager();
         try {
@@ -64,7 +64,7 @@ public class VeiculoRepository {
     }
 
     // remove o veículo com o ID informado; ignorado silenciosamente se não existir
-    public void excluir(Long id) {
+    public void excluir(Integer id) {
         if (id == null || id <= 0) throw new IllegalArgumentException("ID inválido para exclusão.");
         EntityManager em = JPAUtil.getEntityManager();
         try {
